--- timeline/src/main/kotlin/gov/nasa/jpl/aerie/timeline/ops/DirectiveOps.kt
@@ -1,10 +1,0 @@
-package gov.nasa.jpl.aerie.timeline.ops
-
-import gov.nasa.jpl.aerie.timeline.payloads.activities.Directive
-
-/**
- * Operations mixin for timelines of activity directives.
- *
- * Used for both generic directives, and specific directive types from the mission model.
- */
-interface DirectiveOps<A: Any, THIS: DirectiveOps<A, THIS>>: ActivityOps<Directive<A>, THIS>
