+++ timeline/src/main/kotlin/gov/nasa/jpl/aerie/timeline/ops/GeneralOps.kt
@@ -258,7 +258,7 @@ interface GeneralOps<V: IntervalLike<V>, THIS: GeneralOps<V, THIS>>: Timeline<V,
    * @param truncateMarginal whether objects with only partial overlap with an interval in the windows timeline
    *                         should be truncated to the intersection or included unchanged.
    */
-  fun <OTHER: SerialIntervalOps<OTHER>> filterByWindows(windows: SerialIntervalOps<OTHER>, truncateMarginal: Boolean = true) =
+  fun filterByWindows(windows: Windows, truncateMarginal: Boolean = true) =
       if (truncateMarginal) {
         unsafeMap2(ctor, windows) { l, _, i -> l.withNewInterval(i) }
       } else {
